move lsr to own file
